Small reliability improvements over database connection code in all apps. Set the times to default against GMT instead of server's local time. Making mail app. - Added signature support. - Draft can be saved and resumed. - Reliability improvements. - Added preference option of HTML or plain text upon reading mail. Dropped contributor section from meta.xml to be combined as developers. Split persistent connection setting for user and system database. Other small fixes and improvements.
git-svn-id: http://dev.deskportal.net/source/live@32 ffa8ddc6-ed64-449d-89c4-29ad718b3e2b

+++ bookmark/1/0/0/server/php/module/group.php
@@ -33,24 +33,20 @@
 			if(!$user->valid) return false;
 
 			$database = $system->database('user', __METHOD__, $user);
-			if(!$database->success) return false;
 
-			$database->begin(); #Keep the category and relation together
+			if(!$database->success) return false;
+			if(!$database->begin()) return false; #Keep the category and relation together
 
 			$query = $database->prepare("DELETE FROM {$database->prefix}category WHERE id = :id AND user = :user");
 			$query->run(array(':id' => $id, ':user' => $user->id));
 
-			if(!$query->success)
-			{
-				$database->rollback();
-				return false;
-			}
+			if(!$query->success) return $database->rollback() && false;
 
 			$query = $database->prepare("DELETE FROM {$database->prefix}relation WHERE user = :user AND category = :category");
 			$query->run(array(':category' => $id, ':user' => $user->id));
 
-			$database->commit();
-			return $query->success;
+			if(!$query->success) return $database->rollback() && false;
+			return $database->commit() || $database->rollback() && false;
 		}
 
 		public static function set($name, $id = 0, System_1_0_0_User $user = null) #Set a category
